Disabling eval in async functions in non-strict mode
If a var declaration occurs in eval in a non-strict async function the
variable should ideally bind to the inner function. But right now if the
var declaration is shadowing one of the formals then the binding goes
wrong. This changelist makes it a SyntaxError to use eval inside async
functions in non-strict mode. Fix for the binding issue will be addressed later.