Simd.js fixes for Tvs Bugs/Warnings
Refactored code to avoid passing pointers to string buffers as function
arguments while generating simd.js type error messages.

Fixes the following Tvs warnings:
7242752
7243156
7243327
7243735
7243803
7244175
7244414
7244451
7245068
7245507

Note: Please note that we are unable to run the 'oacr' tool because of a
permission issue with the sd enlistments. So we couldn't validate that
this fix has addressed all the 'oacr' warnings in simd.js.