OS 2970825: Spread should throw a syntax error with unary operators
This is a bug is around from an early spec draft. Spread should throw a syntax error when it is not immediately preceeded by a comma, LParen, or LBrack. This fixes an assert that happens when we try to do codegen. There is no fre impact as it gets caught by assignment to rvalue.