Implement ICU based platform agnostic interface on Linux builds
This change implements the UnicodeText Platform Agnostic interface on Linux builds.
With this change, Linux builds will temporarily have an ICU dependency
(we'll do another change later to make it optional but it shouldn't be too difficult).

Additionally, one new API is introduced to the UnicodeText interface- ChangeStringCaseInPlace.
This is a thin wrapper around CharUpperBuffer/CharLowerBuff on Win32.
On Linux it just falls back to the ICU based function ChangeStringLinguisticCase.

Note that while the ICU functions are generally analogous to their Windows counterparts, they
need not be identical in behavior. This can cause differences between Windows and Linux but we'll
deal with them as we encounter them.

Finally, one additional function was re-implemented for Linux. LogicalStringCompareImpl is a
reimplementation of CompareStringW(NORM_IGNORECASE | SORT_DIGITSASNUMBERS)