Fixing bug with FromVar hoisting
A combination of math type spec, copy-prop, nested loops and invariant
hoisting is required to repro this bug. We end up hoisting a FromVar from
a region where it doesn't need a bailout to one where it does. We should
then use the value type of the src in the destination region to put the
right bailout on FromVar.

Addressed comments

Added lossy type checks
Added an assert for the expected bailouts on FromVar