Fixing scope issues with function expressions with split scope
This change fixes multiple issues with using function expression symbols in split scope.

One issue was if the function expression is not captured and has not eval
then we were adding the function expression symbol to body scope. In split
scope case I changed it to add to the param scope. But we don't have to
duplicate the function expression name symbol in the body.

Function expression should be pop out only after taking out the param scope. Another issue was while creating the new frame display during BeginBodyScope we have to skip the function expression.

In split scope the body scope was always getting marked as must
instantiate even though no closure was there in the body scope. I added a
new flag to the scope object to mark whether a closure is there.