cross platform: fix stack overflow (ulong)
ulong (unsigned long) is 4 bytes on Windows while 8 bytes on *nix.

There is/are routines within ChakraCore assume casting from `double`
to `ulong` and then reacing to second item `[1]` is safe. However,
this ends up with stack overflow on *nix.

PAL has a definition for uint32 as (unsigned int) and it is already
widely used. Thus, replacing all the `ulong` to `uint32`