Disable stack nested functions for split scope
When we have non-escaping functions in both param and body scope we are
trying to create stack nested functions for them. The problem is we try to
reuse the local frame display. In case of split scope we cannot do that as
both functions should get different frame displays. This changelist
disables stack nested functions inside split scoped functions.