Function.name on the getter/setter on defineProperty.
For getter/setter function in the object literal we were initially giving them get/set name respectively. However if they appear in the defineProperty we were changing them for diagnostics purpose. In the diagnostics side we were expected to give <propname>.get but due to function.name we have changed it to "get <propname>"., which is not needed and not consistent.
This change is removing that and putting the <propname>.get as full name for the current function. The short name will now automatically be given as get/set correctly.