xplat: enable ch [flaglist] and printing assertions
Fix `OnChakraCoreLoadedEntry` export from `ch`.

Fix `GetModuleHandle(nullptr)`. It was returning HMODULE to
"libChakraCore.so". HMODULE to `ch` was expected. Add the implementation
to PAL.

Deprecate `/` as a command line switch start char on xplat. On Unix it
starts a path.

With this change, `ch` now supports [flaglist], and `AssertMsg` starts
to print assertion line/col/message to console (Previously not because
AssertToConsole test hook flag failed to set).