change back to protect pages with NoAccess instead of partial decommit
it's hard to make the partial decommit withou refactoring page allocator, especially in background releasing pages, there will have race issues. changing to protect with NOACCESS for now.
In the future after we implemented the capability of partially commit pages in pageAllocator we can change to use that, because that can save some memories.