Chakra Assert : sym->HasScopeSlot()
CVE-2016-3202

A let/const bug exposes a bad array indexing operation during bytecode
emit.  A symbol can fail to get a location due to this bug but then the
location is used to offset into an array assuming it is always valid.
Fix the array offset issue by performing a bounds check and throwing
FatalInternalError if an out of bounds access would occur.