Fix for pid ref stack issue in split scope when a lambda shadows a param
When a lambda in the param scope has a parameter that shadows a parameter in the parent function an additional pid ref is inserted into the ref stack as we parse the arguments list of lambda initially as an expression. This causes the parent function to be marked as split scope and the symbols require scope slots. But as there are no non-local references slot register was not allocated for the function. The fix is to assign frame slots register when the function has split scope.