Function.name issue when the LHS is not identifier ref.
The variable isNameIdentifierRef is used to guide the logic to give the function.name. That flag will be set when the LHS is reference identifier (say foo = function() {}), that works correct in the no-deferparse case. For example foo.bar =function() {}, the variable isNameIdentifier should be false. Which works correct till the function is deferred. However when we undefer that function, we parse the function after assignment (foo.bar information is lost). in that case the isNameIdentifier was not set correctly. Fixed that by setting back that flag to new function body from the old parseablefunctioninfo, after the parsing is done.
UT passed. Added one test.