Disable ES6 experimental features through compile flag
JsRuntimeAttributeEnableExperimentalFeatures attribute to JsCreateRuntime
enables all experimental ES6 features together. Selective features can be
disabled using compile time options. When a feature is disabled config
takes precedence and it can be enabled.

The scenario is to disable certain feature when chakracore is compiled
with node but still being able to use test/chk build of
Microsoft/ChakraCore to work with node-chakracore for testing.