Improve ChangePrototype scenario
Whenever we change prototype of an object, we invalidate all proto inline caches
of all properties present in the object's prototype chain. However this can be avoided
if the object is not a prototype of any other object since it would have never got
registered in proto inline cache as a prototype object. So, skip invalidation if that is the
case. The only thing that should be done in this case is change the type of object (if it does
not have locked type) so if it's properties were using proto inline caches, they will be
invalidated by type mismatch and we will update their proto inline cache on next property
access.

Added assert to verify that the object is not in prototype chain of any type registered
in inline cache.