Few iterator close - array destructuring.
Array destructuring was wrapped with try..finally as it was assumed that we will get abrupt completion due to throw. But 'yield' makes it possible to abrupt completion with 'return'. Fixed that by extend that to try.catch.finally.
Also we were emitting reference of the LHS after emitting of RHS. However that is supposed to happen first. Fixed that by Emit reference of LHS first.
Added unittest.