cross-platform: fix AutoSystemInfo Initialization
AutoSystemInfo class was being initialized before its cross dependencies are able to initialize.

This is a compiler/linker dependent issue, requires that AutoSystemInfo is initialized
at the end of the initializion phase.

Although the problem at hand  was not necessarily requiring a thread safe solution, the one proposed here
(static function type of initializer) is thread safe and incredibly liteweight. It does the job by
passing the first `__cxx_global_var_init` (or similar) phase for its host module.