Caching the arguments type for strict mode on the library
Premise:
For strict mode, we start with a simple dictionary type for arguments object and transition to dictionary type upon adding setter/getter. This is costly.

Fix:
We, now, cache the dictionary type with accessors on the javascript library - so that we don't want to do the transition everytime.

This improves the strict mode with arguments by 40% on micro benchmarks. Note that this will be evident only when stack arg optimization is disabled (because with optimization - we will not be creating the heap arguments object in the first place).

Perf tests:
Surprisingly, this code change didn't show up evidently in Speedometer - especially given the fact that there are some functions that have unoptimized heap arguments object.
I am yet to try with acme-benchmark.

IE PerfLab: (Few overall improvements in Elapsed/CPU Time)
http://ieperf-web-01/perfresults/Comparison?testRunId=423952&baseRunId=423947&lab=IEPERF&getRelatedRunData=false&priorities=0,1&categories=failures,warnings,improvements,no-change&search=

Note:
SetAccessor API is still hot as most of the time is spent in GetPropertyName() -> which looks up the propertyRecord from the Dictionary. I kinda gave up at this point on further making strict mode scenario faster as the returns wasn't really fruitful. Let me know if someone has any thoughts on this.