Implement fast path for new Array(const)
Fix the original bug where we should be calling
array fast path if there is a single argument to `Array`
ctor which is constant.

After the bug fix, we were setting `FunctionBodyWeakRef` twice for array initializing.
First time in fast-path and next time from helper path. Updated to not
go through the 2nd code path.