Add delay free listing for func arenas in debug builds
Freed memory on func arenas are returned to a delay free list instead of
making them available immediately for allocation. For Opnd frees,
isDeleted bit is set, this enables detecting erroneous code that can use-afer-free of Opnd
memory. At the end of a phase, the delay free list is merged into the free
list making it available for subsequent allocations.