SharedArrayBuffer - Initial work.
This work is to implement a prototype version of the `SharedArrayBuffer`. The spec is in the stage2.
The `SharedArrayBuffer` is behind the `ESSharedArrayBuffer` (or -sab)
flag.
Highlights.
Introduce the `SharedArrayBuffer` type and its implementation.
Refactor the `ArrayBuffer` to a common class (`ArrayBufferBase`) so that both `SharedArrayBuffer` and `ArrayBuffer` leverage the common functionality and machinery.
`Atomics` object (spec'ed) is introduced to provide the atomic operation on the buffer
which is shared. Currently it is using the Win32 based APIs to provide the
functionality.
All 12 methods of `Atomics` are implemented.
All `TypedArray` views are changed to make use of `SharedArrayBuffer` as well.
The `Serialization/Deserialization` implementation is in the different
repo.
Added test cases to validate most of the functionality. sharedarraybuffer - initial work