Statically link C++ Runtime into ChakraCore.dll.
(Request from TypeScript team.)

This enables us to support running applications developed against
ChakraCore.dll even if the C++ Runtime is not installed on the machine.

Add this to Build/Chakra.Build.Default.props to set the flag for the entire
build. If the user wishes to dynamically link the C++ Runtime, build with e.g.

msbuild ... "/p:RuntimeLib=dynamic_library"

(or any other value for RuntimeLib) to disable the static linking.

Confirmed that ch.exe is now able to run a hello.js file on a machine without
the C++ Runtime installed. (Dynamically linked ChakraCore.dll fails on the same
machine complaining of missing DLLs.)