ModuleNamespace Support
Add support for module namespace. The basic syntax is
export * as myNS from 'submodule.js'
The namespace is a sealed object with all the exported properties from the currently
module. The properties are not changable. They are also "reference property" in that
it should look up the original property instead of caching the value of the property.

There are three set of properties in the namespace object:
. builtin Symbol properties, include Symbol.toStringTag & Symbol.iterator at this time.
. local exports. As the local exports are allocated in the specific slot array.
. nonlocal exports. We need to forward the call to other referencing module namespace object.

Similarly, the enumerator needs to take care of the different kinds of properties. However, the
iterator need to iterate through the property names in sorted order. We'll create the iterator
for a sorted name list. CreateListIterator operation (7.4.8) is implemented as well.

Add a set of unittests to cover the object behavior of namespace object.