Abstract usage of _beginthreadex in PlatformAgnostic class
PlatformAgnostic::Thread::Create exports the functionality we use
consistently in Chakra. In particular the "security" and "thread
address" parameters are missing. See the MSDN for details.

The Windows implementation just calls _beginthreadex. On Linux we use
PAL's CreateThread, but in the future we would like to use pthread
directly. Future patches will enable the functionality on !Windows,
for now it remains disabled. Also, bin/ch and the other binaries will
not use this for the moment, since PlatformAgnostic has a dependency
on Runtime.Base as of today.