Add Module Loader runtime objects
Based on the W3C Module Loader spec:
https://whatwg.github.io/loader

Includes more or less complete implementations of the runtime objects:
%Registry%
%RegistryPrototype%
%Loader%
%LoaderPrototype%
%ModuleStatus%
%ModuleStatusPrototype%

Includes a stub implementation of %Module%.

Most of the abstract operations in the spec are implemented completely but
we are missing some key ones such as SatisfyInstance. The next step will
be to merge the existing ModuleRecord implementation with the ModuleStatus
object and set of abstract operations.

In particular, this pull request does not plug-in any of the dependency
graph resolution or export name resolution because a refactor of existing
ModuleRecord code is required.

There are some open issues with the code in this change. For example, it's
not clear from the spec what we should do if transforming a promise object
with a fulfillment handler returns a non-promise object. It seems possible
that user code could replace the constructor property on the promise
instance which could cause the transformed promise to be anything. Should
we throw here? Should we just wrap whatever the object is in a resolved
promise? It's not clear. Most of the machinery in the spec requires
promise objects so I have taken the approach that we should throw anytime
an object is not a promise when we expect it should be.