Fix bug in assert in AddShadowedData()
DictionaryPropertyDescriptor<>::AddShadowedData() has an assert at the
end to verify the data property index is correct after the shadowing.
However GetDataPropertyIndex<>() relies on the PropertyLetConstGlobal
attribute being set on the Attributes field in order to return correct
results and this was being set after the call to AddShadowedData.

Fixed by moving the attribute flag set into AddShadowedData before the
assertion.