Fix incorrect error when calling non-existent symbol property
We threw 'Object does not support ToString' when trying to call a symbol
property on an object when that property didn't exist.  This happened
because we try to do a toString conversion on the index var for the
error message string, but Symbol objects do not support toString via
JavascriptConversion::ToString().

Fix it by special casing symbol properties and directly calling the
static JavascriptSymbol::ToString method to get a useful string.

Fixes #1409