OSX SDK 10.12 build break
Just saw this after updating one of my dev machines.
CC is not compiling on a machine with Xcode 8.0+ installed due to changes to latest clang.

`clang: warning: libstdc++ is deprecated; move to libc++ with a minimum deployment target of OS X 10.9`

This warning breakes the CMake feature detection scripts.

We could update to libc++. However, that would break node-chakracore.
In order not to break node-chakracore builds,
I'm turning off the cxx `-werror` for osx until node-chakracore moves to libc++